Way to find the constants Cp, Cpp and Ce
In this code i write the decay width as a sum of 6 factors depending on which mix of constants are next to them. Then I compute the terms for each transitions and integrate it for the dipion mass^2. This numerical values are used in a mathematica code to obtain the Cp, Cpp and Ce values using the transitions 5s-3s, 4s-2s and 3s-2s of bottom.